SA-0MM36ESPX09ZLRG1: Extract scheduler types and utilities to dedicated module
Create ampa/scheduler_types.py containing CommandSpec, SchedulerConfig,
RunResult, CommandRunResult dataclasses and shared utility functions
(_utc_now, _to_iso, _from_iso, _seconds_between, _bool_meta).

- scheduler.py: removed inline definitions, imports and re-exports from
  scheduler_types for backward compatibility
- scheduler_cli.py: types now imported from scheduler_types
- delegation.py: replaced local _utc_now/_from_iso copies and deferred
  _bool_meta/CommandRunResult/RunResult imports with module-level imports
  from scheduler_types
- triage_audit.py: replaced local _utc_now/_to_iso/_from_iso copies with
  imports from scheduler_types
- 17 test files: updated to import types from ampa.scheduler_types

All 897 tests pass. Purely structural refactoring with no behavioral changes.
