+++ tests/test_hung_child_recovery.py
@@ -22,16 +22,18 @@ from unittest import mock
 
 import pytest
 
-from ampa.scheduler import (
+from ampa.scheduler_types import (
     CommandRunResult,
     CommandSpec,
     RunResult,
+    _utc_now,
+    _to_iso,
+)
+from ampa.scheduler import (
     Scheduler,
     SchedulerConfig,
     SchedulerStore,
     default_executor,
-    _utc_now,
-    _to_iso,
 )
