+++ tests/test_scheduler_list.py
@@ -1,6 +1,7 @@
 import datetime as dt
 
-from ampa.scheduler import SchedulerStore, CommandSpec
+from ampa.scheduler_types import CommandSpec
+from ampa.scheduler import SchedulerStore
 from ampa.scheduler_cli import _build_command_listing
