+++ tests/test_scheduler_scoring.py
@@ -2,11 +2,13 @@ import datetime as dt
 import json
 import os
 
-from ampa.scheduler import (
+from ampa.scheduler_types import (
     CommandSpec,
     RunResult,
-    Scheduler,
     SchedulerConfig,
+)
+from ampa.scheduler import (
+    Scheduler,
     SchedulerStore,
 )
